Refactor and simplify kmsg logging
This is possible now that syslog logging is completely separate.
