+++ config/config.exs
@@ -1,6 +1,4 @@
 use Mix.Config
 
-config :nerves_runtime, :enable_syslog, Mix.env() != :test
-
 config :nerves_runtime,
   target: "host"
